ml_pipeline/show_visualize_deeplake/requirements.txt
deeplake==3.6.14
pillow==9.4.0
numpy==1.22.4
